ft_pipeline/finguardium-leviosa/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.1
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distro==1.9.0
dnspython==2.6.1
Flask==3.0.3
Flask-RESTful==0.3.10
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.28.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
validators==0.28.1
weaviate-client==3.26.2
Werkzeug==3.0.3
